+++ lib/src/flutter/utilities.dart
@@ -1,5 +1,0 @@
-export 'utilities/padding.dart';
-
-export 'utilities/radius.dart';
-
-export 'utilities/colors.dart';